# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HUGAMEX API",
    "description": "Backend-only corporate content, authentication and RBAC CMS. Cookie auth transitions require X-CSRF-TOKEN; admin operations use Bearer access tokens and role/MFA checks.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/auth/verify-reset-otp": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "verifyReset",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "verify",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "reset",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "resend",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOnly"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "refresh",
        "parameters": [
          {
            "name": "HUGAMEX_REFRESH",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "tags": [
          "mfa-controller"
        ],
        "operationId": "verify_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mfa"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/setup": {
      "post": {
        "tags": [
          "mfa-controller"
        ],
        "operationId": "setup",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/recovery": {
      "post": {
        "tags": [
          "mfa-controller"
        ],
        "operationId": "recover",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mfa"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/recovery/regenerate": {
      "post": {
        "tags": [
          "mfa-controller"
        ],
        "operationId": "regenerate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/enable": {
      "post": {
        "tags": [
          "mfa-controller"
        ],
        "operationId": "enable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mfa"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "logout",
        "parameters": [
          {
            "name": "HUGAMEX_REFRESH",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "forgot",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "GET /api/v1/auth/csrf first; retain its cookie and send the returned token here.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOnly"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/account/change-password": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Change"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/csrf": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "csrf",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CsrfToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/config": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "config",
        "parameters": [
          {
            "name": "HUGAMEX_REFRESH",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Verify": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "pattern": "[0-9]{6}"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
      "Reset": {
        "type": "object",
        "properties": {
          "resetToken": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          }
        },
        "required": [
          "password",
          "resetToken"
        ]
      },
      "EmailOnly": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          }
        },
        "required": [
          "email"
        ]
      },
      "Challenge": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Register": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "Mfa": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          }
        },
        "required": [
          "code"
        ]
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "Change": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          }
        },
        "required": [
          "currentPassword",
          "password"
        ]
      },
...
      "CsrfToken": {
        "type": "object",
        "properties": {
          "headerName": {
            "type": "string"
          },
          "parameterName": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}